Add support for field aliases and multiple root fields
It is now possible for a query to contain more than root field as
described in the spec discussion linked to in #13. Where the root
queries differ only in their arguments they must be renamed using
field aliases (described in section 2.7 of the spec) to ensure that
their result fields don't conflict.

Adding the support for renaming revealed a bug in the schema/interpreter
composition mechanism which is now fixed.

Adding support for multiple root nodes revealed a bug in the handling of
Unique queries where the expected type is nullable. Prior to this
commit, if there was no unique match the query would fail incorrectly
rather than returning a null value. This is now also fixed.