Fixed issue with nested lists
Nested list structures which generated SQL unions on the left hand side
of a join could cause alias confusions leading to either incorrect SQL
being generated or, if the SQL was correct, incorrect indexing into the
result set.

This PR fixes that and also simplified and cleans up the the nest
operation on SQLUnion.