Make Like polymorphic in the nullability of its Term
The SQL LIKE operator can be applied to both nullable and non-nullable
columns, however, prior to this commit, Grackle's Like predicate is only
applicatble to non-nullable Terms.

This PR allows Like to be applied to both nullable and non-nullable
Terms. For Scala 3, this is captured direction using a union type for
the Term (Term[String]|Term[Option[String]]). For Scala 2 we need a
slightly more encoding using an implicit type constraint to ensure
correctness outside the sql module, and a weaker type signature
(Term[_]) internally.

A new SqlLike test suite has been added, which illustrates a plausiable
scenario where this might be desirable (as opposed to simply typing the
term as Term[String] in all cases). In this example method mkPredicate
constructs a filter which applies a Like where a pattern is provided, or
applies a null test when no pattern is provided. This method would be
quite awkward to write if Like could only take a Term[String] as an
argument because there is no straightforward way of obtaining a
Term[String] from a Term[Option[String]] or vice versa.