Rename the doobie module to doobie-pg
Both the skunk and the doobie module share the logic of the sql module
which by now generates very Postgres-specific SQL. At some point we will
almost certainly want to support other DBs via doobie, but right now the
doobie module in grackle isn't as generic as it's name implies.

To avoid painting ourselves into a corner the PR renames the doobie
module to doobie-pg and changes the package name to doobie.postgres.