Embedding fixes
When subobjects are embedded in an enclosing select rather than via a
join we have to ensure that all the corresponding context paths align.
This has been done by wrapping the nested columns with EmbeddedColumn
which preserves the embedded path of the underlying column whilst
simultaneously matching the path of the enclosing select.

However a number of cases where this needed to be done (references in
where and order clauses) were missed and also the generation of columns
for parent constraints didn't take embedding into account.