Cursors resolve fields internal before searching other mappings
Previously cursors attempt to resolve fields globally, potentially
delegating to an alternative mapping, before resolving relative to the
current cursor. In the case of the circe and generic mappings this could
result in other mappings hijacking field resolution even when the
currently focussed values could resolve the field successfully. Whilst
this behaviour wasn't completely unmotivated, it runs counter to the way
Grackle mappings are composed in practice.